requirements.txt
Flask~=2.2.2
flask-cors
Werkzeug~=2.2.2
wheel~=0.38.4
redis
rq
asyncio
python-dotenv
boto3
pyjwt
GPUtil